README.md
## Important note ##

This repository does not include backend or smart contracts since our team was made up of two non-technical members. During the hackathon, we focused on conceptualization, design, and a visual prototype (mainly CSS and basic front-end).

While we couldn’t advance on the technical development, we built the full vision of the project and are confident about its social impact and value. Our next step is to bring developers on board to take this idea into production.

We strongly believe in the potential of this platform and will continue working to make it a reality.